% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{add_lake_item}
\alias{add_lake_item}
\title{Add lake item}
\usage{
add_lake_item(body, credentials)
}
\arguments{
\item{body}{The body of the request as json data (use `jsonlite::toJSON(body, auto_unbox=T)`).}

\item{credentials}{A credentials object to authenticate with.}
}
\description{
Add lake item
}
