% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{add_table_item}
\alias{add_table_item}
\title{Add table item}
\usage{
add_table_item(table_name, body, credentials)
}
\arguments{
\item{body}{The body of the request, as a named list.}

\item{credentials}{A credentials object to authenticate with.}
}
\description{
Add table item
}
