% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{batch_create_table_items}
\alias{batch_create_table_items}
\title{Batch create table items}
\usage{
batch_create_table_items(table_name, body, credentials)
}
\arguments{
\item{body}{List of json elements.}

\item{credentials}{A credentials object to authenticate with.}
}
\description{
Batch create table items
}
