% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_version}
\alias{check_version}
\title{Check if package version is up to date}
\usage{
check_version(warn_only = F)
}
\arguments{
\item{warn_only}{Logical. TRUE to only return warning if the package version is not up to date. FALSE (default) to return a message confirming package is up to date.}
}
\description{
Check if package version is up to date
}
