% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create}
\alias{create}
\title{Create}
\usage{
create(path, body, credentials)
}
\arguments{
\item{path}{The path to the API endpoint, excluding the hub URL (and not starting with a forward slash).}

\item{body}{The body of the request, as a named list.}

\item{credentials}{A credentials object to authenticate with.}
}
\description{
Create
}
