% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{filter_previous}
\alias{filter_previous}
\title{Filter previous}
\usage{
filter_previous(last_result, credentials)
}
\arguments{
\item{last_result}{Last result.}

\item{credentials}{A credentials object to authenticate with.}
}
\description{
Filter previous
}
