% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{filter_table_items}
\alias{filter_table_items}
\title{Filter table items}
\usage{
filter_table_items(table_name, credentials, filter)
}
\arguments{
\item{credentials}{A credentials object to authenticate with.}
}
\description{
Filter table items
}
