% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{filter_tables}
\alias{filter_tables}
\title{Filter tables}
\usage{
filter_tables(credentials, filter)
}
\arguments{
\item{credentials}{A credentials object to authenticate with.}
}
\description{
Filter tables
}
