% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{form_post}
\alias{form_post}
\title{Call an API using the POST method as a multipart/form-data request}
\usage{
form_post(path, body, credentials = NULL, verify = T, timeout = 30)
}
\arguments{
\item{path}{The path to the API endpoint, excluding the hub URL (and not starting with a forward slash).}

\item{body}{The body of the request as json data (use `jsonlite::toJSON(body, auto_unbox=T)`).}

\item{credentials}{A credentials object to authenticate with.}

\item{verify}{Optional. Whether to verify the SSL certificate of the API. Defaults to TRUE.}

\item{timeout}{Optional. The timeout in seconds to wait for a response. Defaults to 30.}
}
\description{
Call an API using the POST method as a multipart/form-data request
}
