% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get}
\alias{get}
\title{Call an API using the GET method}
\usage{
get(path, options = NULL, credentials = NULL, verify = T, timeout = 30)
}
\arguments{
\item{path}{The path to the API endpoint, excluding the hub URL (and not starting with a forward slash).}

\item{options}{Optional list of query parameters in a named list}

\item{credentials}{A credentials object to authenticate with.}

\item{verify}{Optional. Whether to verify the SSL certificate of the API. Defaults to TRUE.}

\item{timeout}{Optional. The timeout in seconds to wait for a response. Defaults to 30.}
}
\description{
Call an API using the GET method
}
