% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_credentials}
\alias{get_credentials}
\title{Generate a credentials object to pass to this package's functions in order to authenticate to the API.
If either the username or password are not specified, they will be asked interactively.}
\usage{
get_credentials(hub_url, username = NULL, password = NULL)
}
\arguments{
\item{hub_url}{The URL of the hub to authenticate to. Must end with a forward slash.}

\item{username}{Optional. The username to authenticate with.}

\item{password}{Opional. The password to authenticate with.}
}
\description{
Generate a credentials object to pass to this package's functions in order to authenticate to the API.
If either the username or password are not specified, they will be asked interactively.
}
