% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_empty_credentials}
\alias{get_empty_credentials}
\title{Create an empty credentials object. Useful for internal use or alternative login methods.
This is generally not useful for general users.}
\usage{
get_empty_credentials(hub_url)
}
\arguments{
\item{hub_url}{The URL of the hub to authenticate to. Must end with a forward slash.}
}
\description{
Create an empty credentials object. Useful for internal use or alternative login methods.
This is generally not useful for general users.
}
