% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list_}
\alias{list_}
\title{List all elements of a specific endpoint}
\usage{
list_(path, credentials)
}
\arguments{
\item{path}{The path to the API endpoint, excluding the hub URL (and not starting with a forward slash).}

\item{credentials}{A credentials object to authenticate with.}
}
\value{
depends on the endpoint. Usually a named list with a "results" key containing a list of elements, or a list of elements.
}
\description{
The underscore is to prevent the function for overwriting the base list() function
}
