% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{list_lake_items}
\alias{list_lake_items}
\title{List lake items}
\usage{
list_lake_items(credentials)
}
\arguments{
\item{credentials}{A credentials object to authenticate with.}
}
\description{
List lake items
}
