% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list_paginated}
\alias{list_paginated}
\title{Get the first set of a paginated list of elements and the information to get the following pages.}
\usage{
list_paginated(path, credentials, cursor = NULL)
}
\arguments{
\item{path}{The path to the API endpoint, excluding the hub URL (and not starting with a forward slash).}

\item{credentials}{A credentials object to authenticate with.}

\item{cursor}{Optional. The cursor to use for pagination. Defaults to NULL}
}
\description{
Get the first set of a paginated list of elements and the information to get the following pages.
}
