% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list_previous}
\alias{list_previous}
\title{List previous results}
\usage{
list_previous(last_result, credentials)
}
\arguments{
\item{last_result}{Last result.}

\item{credentials}{A credentials object to authenticate with.}
}
\description{
List previous results
}
