% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{list_table_items}
\alias{list_table_items}
\title{List table items}
\usage{
list_table_items(table_name, credentials, cursor = NULL)
}
\arguments{
\item{credentials}{A credentials object to authenticate with.}

\item{cursor}{Optional. The cursor to use for pagination. Defaults to NULL}
}
\description{
List table items
}
