% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{options}
\alias{options}
\title{Call an API using the OPTIONS method}
\usage{
options(path, credentials = NULL, verify = T, timeout = 30)
}
\arguments{
\item{path}{The path to the API endpoint, excluding the hub URL (and not starting with a forward slash).}

\item{credentials}{A credentials object to authenticate with.}

\item{verify}{Optional. Whether to verify the SSL certificate of the API. Defaults to TRUE.}

\item{timeout}{Optional. The timeout in seconds to wait for a response. Defaults to 30.}
}
\description{
Call an API using the OPTIONS method
}
