% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove}
\alias{remove}
\title{Remove}
\usage{
remove(path, credentials)
}
\arguments{
\item{path}{The path to the API endpoint, excluding the hub URL (and not starting with a forward slash).}

\item{credentials}{A credentials object to authenticate with.}
}
\description{
Remove
}
