% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{remove_lake_item}
\alias{remove_lake_item}
\title{Remove lake item}
\usage{
remove_lake_item(id, credentials)
}
\arguments{
\item{credentials}{A credentials object to authenticate with.}
}
\description{
Remove lake item
}
