% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{retrieve_journal}
\alias{retrieve_journal}
\title{Retrieve journal}
\usage{
retrieve_journal(id, credentials)
}
\arguments{
\item{credentials}{A credentials object to authenticate with.}
}
\description{
Retrieve journal
}
