% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{update_table_item}
\alias{update_table_item}
\title{Update table item}
\usage{
update_table_item(table_name, id, body, credentials)
}
\arguments{
\item{body}{The body of the request as json data (use `jsonlite::toJSON(body, auto_unbox=T)`).}

\item{credentials}{A credentials object to authenticate with.}
}
\description{
Update table item
}
