% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{batch_create_table_items}
\alias{batch_create_table_items}
\title{The body must be a list of json elements}
\usage{
batch_create_table_items(table_name, body, credentials)
}
\description{
The body must be a list of json elements
}
