% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1.R
\name{batch_delete_table_items}
\alias{batch_delete_table_items}
\title{Batch delete table items}
\usage{
batch_delete_table_items(table_name, body, credentials)
}
\arguments{
\item{body}{List of ids (not keys).}
}
\description{
Batch delete table items
}
